Migrate services

+++ aspnet-core/src/WatchWord.Core/Domain/Entities/Material.cs
@@ -1,5 +1,5 @@
 ﻿using System.Collections.Generic;
-using WatchWord.Domain.Entities.Common;
+using Abp.Domain.Entities;
 
 namespace WatchWord.Domain.Entities
 {
@@ -13,7 +13,7 @@ namespace WatchWord.Domain.Entities
     }
 
     /// <summary>The Material entity, a film or series with information about it.</summary>
-    public class Material : Entity<int>
+    public class Material : Entity<long>
     {
         /// <summary>Gets or sets the material type.</summary>
         public MaterialType Type { get; set; }
